+++ ColorToolkit/Forms/FormMain.cs
@@ -3,8 +3,12 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Drawing;
+using System.Drawing.Pictograms;
+using System.Linq;
 using System.Text;
 using System.Windows.Forms;
+using System.Windows.Forms.Pictograms;
+using Platform.Support.Drawing;
 
 namespace Toolkit.Forms
 {
@@ -17,27 +21,47 @@ namespace Toolkit.Forms
 
             Icon = Icon.ExtractAssociatedIcon(System.Reflection.Assembly.GetEntryAssembly().Location);
 
+            toolStripButtonTopMost.Checked = TopMost;
+
+            toolStripButtonOpen.SetImage(MaterialDesign.Instance, MaterialDesign.IconType.folder_open, 48, Color.White);
+            toolStripButtonQSwatch.SetImage(MaterialDesign.Instance, MaterialDesign.IconType.color_lens, 48, Color.White);
+            toolStripButtonCopy.SetImage(MaterialDesign.Instance, MaterialDesign.IconType.content_copy, 48, Color.White);
+            toolStripButtonHistory.SetImage(MaterialDesign.Instance, MaterialDesign.IconType.restore, 48, Color.White);
+
+            toolStripButtonTopMost.SetImage(MaterialDesign.Instance, MaterialDesign.IconType.layers, 48, Color.White);
+            toolStripButtonAbout.SetImage(MaterialDesign.Instance, MaterialDesign.IconType.info, 48, Color.White);
+            toolStripButtonClose.SetImage(MaterialDesign.Instance, MaterialDesign.IconType.close, 48, Color.White);
+
+            panelColor.BackgroundImage = MaterialDesign.GetImage(MaterialDesign.IconType.colorize, 48, Color.White);
+
+            //copyToolStripMenuItem.SetImage(MaterialDesign.Instance, MaterialDesign.IconType.content_copy, 16, toolStripMenu.BackColor);
+            //pasteToolStripMenuItem.SetImage(MaterialDesign.Instance, MaterialDesign.IconType.content_paste, 16, toolStripMenu.BackColor);
 #if DEBUG
-            this.Color = Helpers.FromHEX("#F44336");
+
+            Color = ColorHelpers.ToColor("#F44336");
             setColor();
+
+            FormHelpers.ExtractResources(panelColor.BackgroundImage, panelColor.Name);
+            FormHelpers.ExtractResources(toolStripMenu);
 #endif
+
         }
 
         protected override void OnActivated(EventArgs e)
         {
             base.OnActivated(e);
-            this.Opacity = 1.0;
+            Opacity = 1.0;
         }
         protected override void OnDeactivate(EventArgs e)
         {
             try
             {
-                this.Opacity = 0.25;
+                Opacity = 0.25;
             }
             catch (Exception)
             {
             }
-            
+
             base.OnDeactivate(e);
         }
 
@@ -48,37 +72,37 @@ namespace Toolkit.Forms
             MouseButtons button = e.Button;
             if (button == MouseButtons.Left)
             {
-                this.Cursor = Cursors.Cross;
+                Cursor = Cursors.Cross;
             }
         }
         private void panelColor_DoubleClick(object sender, EventArgs e)
         {
-            this.colorDialogPicker.Color = this.Color;
-            if (this.colorDialogPicker.ShowDialog() == DialogResult.OK)
+            colorDialogPicker.Color = Color;
+            if (colorDialogPicker.ShowDialog() == DialogResult.OK)
             {
-                this.Color = this.colorDialogPicker.Color;
-                this.setColor();
+                Color = colorDialogPicker.Color;
+                setColor();
             }
         }
         private void panelColor_MouseDown(object sender, MouseEventArgs e)
         {
             if (e.Button == MouseButtons.Left)
             {
-                this.Cursor = Cursors.Cross;
+                Cursor = Cursors.Cross;
             }
         }
         private void panelColor_MouseMove(object sender, MouseEventArgs e)
         {
             if (e.Button == MouseButtons.Left)
             {
-                this.getScreenColor();
+                getScreenColor();
             }
         }
         private void panelColor_MouseUp(object sender, MouseEventArgs e)
         {
             if (e.Button == MouseButtons.Left)
             {
-                this.Cursor = Cursors.Default;
+                Cursor = Cursors.Default;
             }
         }
 
@@ -86,29 +110,29 @@ namespace Toolkit.Forms
         {
             if (e.KeyCode == Keys.Return)
             {
-                this.buttonCopyHEX.PerformClick();
+                copyHEXToolStripMenuItem.PerformClick();
             }
             else
             {
                 if (e.KeyCode == Keys.Escape)
                 {
-                    this.textBoxHEX.Text = string.Empty;
+                    textBoxHEX.Text = string.Empty;
                 }
             }
         }
         private void textBoxHEX_Leave(object sender, EventArgs e)
         {
 
-            if (!this.textBoxHEX.Text.StartsWith("#")) { this.textBoxHEX.Text = "#" + this.textBoxHEX.Text.Trim(); }
+            if (!textBoxHEX.Text.StartsWith("#")) { textBoxHEX.Text = "#" + textBoxHEX.Text.Trim(); }
 
-            if (!System.Text.RegularExpressions.Regex.IsMatch("#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})", this.textBoxHEX.Text, System.Text.RegularExpressions.RegexOptions.IgnoreCase))
+            if (!System.Text.RegularExpressions.Regex.IsMatch("#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})", textBoxHEX.Text, System.Text.RegularExpressions.RegexOptions.IgnoreCase))
             {
-                this.Color = Helpers.FromHEX(this.textBoxHEX.Text);
-                this.setColor();
+                Color = ColorHelpers.ToColor(textBoxHEX.Text);
+                setColor();
             }
             else
             {
-                this.textBoxHEX.Text = Helpers.ToHEX(System.Drawing.Color.FromArgb((int)this.textBoxR.Value, (int)this.textBoxG.Value, (int)this.textBoxB.Value));
+                textBoxHEX.Text = ColorHelpers.ToHEX(System.Drawing.Color.FromArgb((int)textBoxR.Value, (int)textBoxG.Value, (int)textBoxB.Value));
             }
 
         }
@@ -117,36 +141,75 @@ namespace Toolkit.Forms
         {
             if (e.KeyCode == Keys.Return)
             {
-                this.buttonCopyRGB.PerformClick();
+                copyRGBToolStripMenuItem.PerformClick();
             }
             else
             {
                 if (e.KeyCode == Keys.Escape)
                 {
-                    this.textBoxR.Value = 0;
-                    this.textBoxG.Value = 0;
-                    this.textBoxB.Value = 0;
+                    textBoxR.Value = 0;
+                    textBoxG.Value = 0;
+                    textBoxB.Value = 0;
                 }
             }
         }
         private void textBoxRGB_Leave(object sender, EventArgs e)
         {
-            this.Color = Color.FromArgb(255, int.Parse(this.textBoxR.Value.ToString()), int.Parse(this.textBoxG.Value.ToString()), int.Parse(this.textBoxB.Value.ToString()));
-            this.setColor();
+            Color = Color.FromArgb(255, int.Parse(textBoxR.Value.ToString()), int.Parse(textBoxG.Value.ToString()), int.Parse(textBoxB.Value.ToString()));
+            setColor();
         }
 
-        private void buttonCopiarHEX_Click(object sender, EventArgs e)
+        private void textBoxHSB_KeyDown(object sender, KeyEventArgs e)
         {
-            Clipboard.SetText(this.textBoxHEX.Text);
+            if (e.KeyCode == Keys.Return)
+            {
+                copyHSBToolStripMenuItem.PerformClick();
+            }
+            else
+            {
+                if (e.KeyCode == Keys.Escape)
+                {
+                    textBoxH.Value = 0;
+                    textBoxS.Value = 0;
+                    textBoxV.Value = 0;
+                }
+            }
         }
-        private void buttonCopiarRGB_Click(object sender, EventArgs e)
+        private void textBoxHSB_Leave(object sender, EventArgs e)
         {
-            Clipboard.SetText(String.Join(",", new string[] { this.textBoxR.Value.ToString(), this.textBoxG.Value.ToString(), this.textBoxB.Value.ToString() }));
+            Color = ColorHelpers.ToColor(new HSB((int)textBoxH.Value, (int)textBoxS.Value, (int)textBoxV.Value));
+            setColor();
+        }
+
+        private void buttonCopyHEX_Click(object sender, EventArgs e)
+        {
+            Clipboard.SetText(textBoxHEX.Text);
+        }
+        private void buttonCopyRGB_Click(object sender, EventArgs e)
+        {
+            Clipboard.SetText(String.Join(",", new string[] { textBoxR.Value.ToString(), textBoxG.Value.ToString(), textBoxB.Value.ToString() }));
+        }
+        private void buttonCopyHSB_Click(object sender, EventArgs e)
+        {
+            Clipboard.SetText(String.Join(",", new string[] { textBoxH.Value.ToString(), textBoxS.Value.ToString(), textBoxV.Value.ToString() }));
+        }
+
+        private void labelHEX_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
+        {
+            buttonCopyHEX_Click(sender, e);
+        }
+        private void labelRGB_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
+        {
+            buttonCopyRGB_Click(sender, e);
+        }
+        private void labelHSB_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
+        {
+            buttonCopyHSB_Click(sender, e);
         }
 
         private void getScreenColor()
         {
-            Bitmap bitmap = new Bitmap(Program.GetWorkingArea().Width, Program.GetWorkingArea().Height);
+            Bitmap bitmap = new Bitmap(FormHelpers.GetWorkingArea().Width, FormHelpers.GetWorkingArea().Height);
             Point mousePosition = Control.MousePosition;
             checked
             {
@@ -159,76 +222,120 @@ namespace Toolkit.Forms
                     Point upperLeftDestination = new Point(0, 0);
                     arg_A4_0.CopyFromScreen(arg_A4_1, upperLeftDestination, bitmap.Size);
                 }
-                this.Color = bitmap.GetPixel((int)Math.Round((double)bitmap.Size.Width / 2.0), (int)Math.Round((double)bitmap.Size.Height / 2.0));
-                this.setColor();
-                this.Invalidate();
+                Color = bitmap.GetPixel((int)Math.Round((double)bitmap.Size.Width / 2.0), (int)Math.Round((double)bitmap.Size.Height / 2.0));
+                setColor();
+                Invalidate();
             }
         }
         private void setColor()
         {
-            this.panelColor.BackColor = this.Color;
-            this.textBoxHEX.Text = Helpers.ToHEX(this.Color);
 
-            this.textBoxR.Value = this.Color.R;
-            this.textBoxG.Value = this.Color.G;
-            this.textBoxB.Value = this.Color.B;
+            if (Properties.Settings.Default.History == null)
+                Properties.Settings.Default.History = new System.Collections.Specialized.StringCollection();
+
+            if (!Properties.Settings.Default.History.Contains(Color.ToArgb().ToString()))
+            {
+                Properties.Settings.Default.History.Add(Color.ToArgb().ToString());
+                Properties.Settings.Default.Save();
+            }
+
+            panelColor.BackColor = Color;
+            panelColor.BackgroundImage = MaterialDesign.GetImage(MaterialDesign.IconType.colorize, 48, Color.Invert());
+
+            toolStripButtonHistory.Enabled = Properties.Settings.Default.History != null && Properties.Settings.Default.History.Count > 0;
+
+            textBoxHEX.Text = ColorHelpers.ToHEX(Color);
+
+            textBoxR.Value = Color.R;
+            textBoxG.Value = Color.G;
+            textBoxB.Value = Color.B;
+
+            var hsv = ColorHelpers.ToHSB(Color);
+
+            textBoxH.Value = (decimal)hsv.Hue360;
+            textBoxS.Value = (decimal)hsv.Saturation100;
+            textBoxV.Value = (decimal)hsv.Brightness100;
+
         }
 
         private void buttonQSwatch_Click(object sender, EventArgs e)
         {
-            FormQSwatch _FormQSwatch = new FormQSwatch(this.Color);
-            _FormQSwatch.Show();
+            displayChild(new FormQSwatch(Color), 1);
         }
 
-        private void buttonPicture_Click(object sender, EventArgs e)
+        private void openFileDialogMain_FileOk(object sender, CancelEventArgs e)
         {
-            this.openFileDialogPicture.ShowDialog();
-        }
-
-        private void openFileDialogPicture_FileOk(object sender, CancelEventArgs e)
-        {
-            foreach (string item in this.openFileDialogPicture.FileNames)
+            foreach (string item in openFileDialogMain.FileNames)
             {
                 try
                 {
-                    Image file = Image.FromFile(item);
-                    this.Color = Helpers.GetDominantColor(file);
-                    this.setColor();
+                    var images = new string[] { ".bmp", ".gif", ".jpeg", ".jpg", ".png" };
+                    var palettes = new string[] { ".ase", ".aco" };
+                    var file = new System.IO.FileInfo(item);
+                    if (images.Contains(file.Extension))
+                    {
+                        var image = Image.FromFile(item);
+                        Color = ColorHelpers.GetDominantColor(image);
+                        setColor();
+                    }
+                    else if (palettes.Contains(file.Extension))
+                        displayChild(new FormSwatch(item), 3);
+
                 }
                 catch (Exception ex)
                 {
-                    System.Windows.Forms.MessageBox.Show(ex.Message, this.Text, MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                    System.Windows.Forms.MessageBox.Show(ex.Message, Text, MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
                 }
             }
         }
 
-        private void buttonOpenSwatch_Click(object sender, EventArgs e)
+
+        private void toolStripButtonClose_Click(object sender, EventArgs e)
         {
-            this.openFileDialogSwatch.ShowDialog();
+            Close();
         }
 
-        private void openFileDialogSwatch_FileOk(object sender, CancelEventArgs e)
+        private void toolStripButtonAbout_Click(object sender, EventArgs e)
         {
-            foreach (string item in this.openFileDialogSwatch.FileNames)
+            var child = new FormAbout();
+            child.ShowDialog();
+        }
+
+        private void toolStripButtonOpen_Click(object sender, EventArgs e)
+        {
+            openFileDialogMain.ShowDialog();
+        }
+
+        private void toolStripButtonTopMost_Click(object sender, EventArgs e)
+        {
+            TopMost = !TopMost;
+            toolStripButtonTopMost.Checked = TopMost;
+            Properties.Settings.Default.TopMost = TopMost;
+            Properties.Settings.Default.Save();
+        }
+
+
+        private void displayChild(Form child, int factor = 0)
+        {
+            if (factor > 0)
             {
-                try
-                {
-                    FormSwatch _FormSwatch = new FormSwatch(item);
-                    _FormSwatch.Show();
-                }
-                catch (Exception ex)
-                {
-                    System.Windows.Forms.MessageBox.Show(ex.Message, this.Text, MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
-                }
+                child.Width = this.Width * factor;
+                child.Height = this.Height;
             }
+            child.StartPosition = FormStartPosition.Manual;
+            if (this.DesktopLocation.X > (FormHelpers.GetWorkingArea().Width / 2))
+                child.Location = new Point(this.Location.X - child.Width - this.Padding.Left, this.Location.Y);
+            else
+                child.Location = new Point(this.Location.X + this.Width + this.Padding.Left, this.Location.Y);
+            child.Show();
         }
 
-        private void buttonAbout_Click(object sender, EventArgs e)
+        private FormHistory FormHistory;
+
+        private void toolStripButtonHistory_Click(object sender, EventArgs e)
         {
-            FormAbout _FormAbout = new FormAbout();
-            _FormAbout.ShowDialog();
+            if (FormHistory == null || FormHistory.IsDisposed) FormHistory = new FormHistory();
+            displayChild(FormHistory, 3);
         }
-
-
     }
 }