+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.3] - [2016-12-28]
+### Changed
+- UI Refresh
+- History window
+- HSB Color generation
+- Optional TopMost
+
+## [1.0.2] - [2016-05-28]
+### Changed
+- Standardization
+- Minors bugs fixes
+
 ## [1.0.1] - [2016-03-10]
 ### Changed
 - UI minors changes